Added clothing sets you can buy.
Removed some of the outfits that are now accessible as single items in the wardrobe and turned them into sets you can buy. I'll also make sets for lingerie, muggle outfits, ... There is also a new shop screen for them.
This is a wip so there is still some stuff missing!